feature-A001-code-refactoring Containerized the application
Also moved one of the tests into correct folder, cleaned-up TODOs, and updated notes.
