+++ .dockerignore
@@ -1,0 +1,5 @@
+*.log
+*.class
+*.iml
+.idea/
+.git/
