+++ Dockerfile
@@ -1,0 +1,4 @@
+FROM openjdk:21-jdk-slim
+WORKDIR /app
+COPY target/refactoring-java-1.0-SNAPSHOT.jar app.jar
+ENTRYPOINT ["java", "-jar", "app.jar"]
